Restructure x/auth/client
