--- x/auth/client/tx/root.go
@@ -1,15 +1,0 @@
-package tx
-
-import (
-	"github.com/gorilla/mux"
-
-	"github.com/cosmos/cosmos-sdk/client/context"
-)
-
-// register REST routes
-func RegisterTxRoutes(cliCtx context.CLIContext, r *mux.Router) {
-	r.HandleFunc("/txs/{hash}", QueryTxRequestHandlerFn(cliCtx)).Methods("GET")
-	r.HandleFunc("/txs", QueryTxsByTagsRequestHandlerFn(cliCtx)).Methods("GET")
-	r.HandleFunc("/txs", BroadcastTxRequest(cliCtx)).Methods("POST")
-	r.HandleFunc("/txs/encode", EncodeTxRequestHandlerFn(cliCtx)).Methods("POST")
-}
